Extract ResolveCasing to Platform
