+++ src/openrct2/platform/Platform2.h
@@ -39,6 +39,7 @@ namespace Platform
     bool IsPathSeparator(char c);
     utf8* GetAbsolutePath(utf8* buffer, size_t bufferSize, const utf8* relativePath);
     uint64_t GetLastModified(const std::string& path);
+    std::string ResolveCasing(const std::string& path, bool fileExists);
     rct2_time GetTimeLocal();
     rct2_date GetDateLocal();
     bool FindApp(const std::string& app, std::string* output);
